refactor : add goNextPage Component
